ref(tech): make suspense management on feature level
Ref: #19
