+++ src/features/checklist/index.ts
@@ -1,3 +1,5 @@
+import CheckList from "./CheckList.vue";
 import CheckListView from "./CheckListView.vue";
 
+export { CheckList };
 export default CheckListView;
