+++ src/layouts/BaseLayout.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
 import { RouterView, useRoute } from "vue-router";
-import * as t from "@locales/messages";
 import { setLanguageTag } from "@locales/runtime";
 
 const { params } = useRoute("checklist");
@@ -10,17 +9,5 @@ setLanguageTag(locale);
 </script>
 
 <template>
-  <RouterView v-slot="{ Component }">
-    <template v-if="Component">
-      <Suspense>
-        <component :is="Component"></component>
-
-        <template #fallback
-          ><section class="flex items-center justify-center text-3xl font-bold">
-            <div class="capitalize">{{ t.pending_indicator_text() }}...</div>
-          </section></template
-        >
-      </Suspense>
-    </template>
-  </RouterView>
+  <RouterView />
 </template>
